requirements.txt
absl-py==0.15.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2016.9.26
cffi==1.15.1
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
contextvars==2.4
cx-Freeze==6.13.2
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==2.10.0
html5lib==0.9999999
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
json5==0.9.11
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
Keras==2.2.1
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==3.3.7
MarkupSafe==2.0.1
mistune==0.8.4
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
nltk==3.6.7
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.5.4
Send2Trash==1.8.0
sendgrid==6.10.0
simplegeneric==0.8.1
six==1.16.0
sniffio==1.2.0
starkbank-ecdsa==2.2.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.5.0
tensorflow-estimator==2.6.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing-extensions==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
